--- domain/src/main/java/com/example/domain/model/Payment.kt
@@ -1,5 +1,0 @@
-package com.example.domain.model
-
-enum class Payment {
-    ON_SITE
-}
