+++ domain/src/main/java/com/example/domain/model/model/Movie.kt
@@ -1,4 +1,4 @@
-package com.example.domain.model
+package com.example.domain.model.model
 
 import java.time.LocalDate
