+++ domain/src/main/java/com/example/domain/model/model/Payment.kt
@@ -1,0 +1,5 @@
+package com.example.domain.model.model
+
+enum class Payment {
+    ON_SITE
+}
